move messaging and node to channel-provider module
