+++ modules/channel-provider/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./channelProvider";
+export * from "./messaging";
+export * from "./node";
