+++ modules/client/src/types.ts
@@ -1,27 +1,16 @@
 import {
-  Address,
   AppRegistry,
   Contract,
   IChannelProvider,
   IChannelSigner,
   IClientStore,
   ILoggerService,
-  IMessagingService,
   INodeApiClient,
   Network,
   NodeResponses,
+  IMessagingService,
 } from "@connext/types";
-import { MessagingService } from "@connext/messaging";
 import { JsonRpcProvider } from "ethers/providers";
-
-export interface NodeInitializationParameters {
-  nodeUrl: string;
-  messaging: IMessagingService;
-  logger?: ILoggerService;
-  userIdentifier?: Address;
-  nodeIdentifier?: Address;
-  channelProvider?: IChannelProvider;
-}
 
 export type InternalClientOptions = {
   appRegistry: AppRegistry;
@@ -29,7 +18,7 @@ export type InternalClientOptions = {
   config: NodeResponses.GetConfig;
   ethProvider: JsonRpcProvider;
   logger: ILoggerService;
-  messaging: MessagingService;
+  messaging: IMessagingService;
   network: Network;
   node: INodeApiClient;
   signer: IChannelSigner;
