+++ modules/types/src/basic.ts
@@ -1,11 +1,7 @@
 import { BigNumber as ethersBN, BigNumberish, Interface, ParamType } from "ethers/utils";
 
 export { Contract } from "ethers";
-export {
-  JsonRpcProvider,
-  TransactionReceipt,
-  TransactionResponse,
-} from "ethers/providers";
+export { JsonRpcProvider, TransactionReceipt, TransactionResponse } from "ethers/providers";
 export { BigNumberish, Network, Transaction } from "ethers/utils";
 
 // special strings
